+++ CHANGELOG.md
@@ -29,6 +29,7 @@ MOC and the Multi-order-map. `MOC.probability_in_multiordermap` has a similar
 behavior but also converts a probability-density into a probability.
 * `STMOC.new_empty()` allows to create a new empty Space-Time MOC.
 * `MOC.from_box` to create rectangular MOCs
+* `MOC.from_astropy_regions` to create MOCs from astropy regions.
 
 ## [0.13.1]